% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHTG.R
\name{readHTG}
\alias{readHTG}
\title{Extract count data from HTG EdgeSeq software's output file}
\usage{
readHTG(data.in)
}
\arguments{
\item{data.in}{The excel file as output from HTG EdgeSeq analysis platform}

\item{path.out}{The output path where a new folder containing the results will be saved (optional).
If not specified, the current working directory is used.}
}
\value{
A folder contaiing two plots and two summary table, one the for positive and the negative control genes.
}
\description{
Takes in the Excel file output from the HTG EdgeSeq analysis software and performs a quality check, normalisation and annotation.
}
\details{
The quality check is performed on 4 positive control and 4 negative control genes.
For the positive controls, samples are considered failures if the percentage of reads allocated to the positive controls is greater than 40%.
For the negative controls, samples are considered failures if the deviation from the expected values is greater than twice the standard deviation accross all samples deviations from the expected value.
}
