% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{qualityCheck}
\alias{qualityCheck}
\title{Plots and summary of Quality Check for an HTG EdgeSeq run}
\usage{
qualityCheck(data.in, TableOut = FALSE)
}
\arguments{
\item{data.in}{The raw count data}

\item{TableOut}{Whether intermediate files will be written
If TRUE, the script will write the clean dataset contaning only the genes of interest}
}
\value{
A folder containing two plots and two summary table, one the for positive and the negative control genes.
}
\description{
Takes in the Excel file output from the HTG EdgeSeq analysis software and performs a quality check, normalisation and annotation.
}
\details{
The quality check is performed on 4 positive control and 4 negative control genes.
For the positive controls, samples are considered failures if the percentage of reads allocated to the positive controls is greater than 40%.
For the negative controls, samples are considered failures if the deviation from the expected values is greater than twice the standard deviation accross all samples deviations from the expected value.
}
