% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHTG.R
\name{readHTG}
\alias{readHTG}
\title{Extract count data from HTG EdgeSeq software's output file}
\usage{
readHTG(data.in, PreFormatted = FALSE)
}
\arguments{
\item{data.in}{The absolute path to the excel file with HTG EdgeSeq analysis platform data}

\item{PreFormatted}{Whether the original file was pre formatted.
If FALSE, the excel is the output from the HTG EdgeSeq analysis platform. This includes a "Data" sheet, a box termed "Sample Names", and the fist positive control should be termed ANT1.
If TRUE, the preformatted excel should contain sample names in the first row should, while the first column should corrspond to gene names.}
}
\value{
A dataframe to be used for other functions in the package. This still includes the positive and negative controls.
}
\description{
Takes in the Excel file output from the HTG EdgeSeq analysis software and prepare table for quality check
}
\details{
The function formats and prepares an input excel file to be used for other functions in the package, mainly qualityCheck()
}
